Backend-For-Trivia/User Authentication and Automated Question Tagging/User Authentication/Get-All-Users/requirements.txt
functions-framework==3.*
Flask==2.0.1
google-cloud-firestore==2.1.3
firebase-admin==5.0.1
